Force float32

+++ dolfyn/adv/motion.py
@@ -408,10 +408,10 @@ def correct_motion(ds,
     ##########
     # Calculate the translational velocity (from the accel):
     ds['velacc'] = xr.DataArray(calcobj.calc_velacc(),
-                                dims=['dirIMU', 'time'])
+                                dims=['dirIMU', 'time']).astype('float32')
     # Copy acclow to the adv-object.
     ds['acclow'] = xr.DataArray(calcobj.acclow,
-                                dims=['dirIMU', 'time'])
+                                dims=['dirIMU', 'time']).astype('float32')
 
     ##########
     # Calculate rotational velocity (from angrt):
